+++ packages/eyes-sdk-core/index.js
@@ -229,6 +229,7 @@ exports.EyesSDK = require('./lib/sdk/EyesSDK')
 exports.makeSDK = require('./lib/sdk/makeSDK')
 
 exports.takeDomSnapshot = require('./lib/utils/takeDomSnapshot')
+exports.takeDomSnapshots = require('./lib/utils/takeDomSnapshots')
 exports.takeDomCapture = require('./lib/utils/takeDomCapture')
 
 exports.EyesRunner = require('./lib/runner/EyesRunner')