+++ packages/eyes-sdk-core/lib/sdk/EyesClassic.js
@@ -696,7 +696,8 @@ class EyesClassic extends EyesCore {
   async tryCaptureDom() {
     try {
       this._logger.verbose('Getting window DOM...')
-      return await takeDomCapture(this._logger, this._context)
+      await EyesUtils.markActiveContext(this._logger, this._context)
+      return await takeDomCapture(this._logger, this._driver.mainContext)
     } catch (ignored) {
       return ''
     }