+++ packages/eyes-sdk-core/lib/server/ServerConnector.js
@@ -859,7 +859,14 @@ class ServerConnector {
   async extractText({screenshotUrl, domUrl, location, region, minMatch, language}) {
     ArgumentGuard.notNull(screenshotUrl, 'screenshotUrl')
     this._logger.verbose(
-      `ServerConnector.extractText called with ${JSON.stringify({screenshotUrl, domUrl, region})}`,
+      `ServerConnector.extractText called with ${JSON.stringify({
+        screenshotUrl,
+        domUrl,
+        region,
+        location,
+        minMatch,
+        language,
+      })}`,
     )
 
     const config = {
@@ -888,14 +895,17 @@ class ServerConnector {
     throw new Error(`ServerConnector.extractText - unexpected status (${response.statusText})`)
   }
 
-  async getEmulatedDevicesSizes() {
+  async getEmulatedDevicesSizes(serviceUrl) {
     this._logger.verbose(`ServerConnector.getEmulatedDevicesSizes`)
 
     const config = {
       name: 'getEmulatedDevicesSizes',
       method: 'GET',
       withApiKey: false,
-      url: GeneralUtils.urlConcat(this._renderingInfo.getServiceUrl(), '/emulated-devices-sizes'),
+      url: GeneralUtils.urlConcat(
+        serviceUrl || this._renderingInfo.getServiceUrl(),
+        '/emulated-devices-sizes',
+      ),
     }
 
     const response = await this._axios.request(config)
@@ -908,13 +918,16 @@ class ServerConnector {
     }
   }
 
-  async getIosDevicesSizes() {
+  async getIosDevicesSizes(serviceUrl) {
     this._logger.verbose(`ServerConnector.getIosDevicesSizes`)
 
     const config = {
       name: 'getIosDevicesSizes',
       method: 'GET',
-      url: GeneralUtils.urlConcat(this._renderingInfo.getServiceUrl(), '/ios-devices-sizes'),
+      url: GeneralUtils.urlConcat(
+        serviceUrl || this._renderingInfo.getServiceUrl(),
+        '/ios-devices-sizes',
+      ),
     }
 
     const response = await this._axios.request(config)