+++ packages/sdk-shared/src/get-test-info.js
@@ -19,7 +19,23 @@ async function getDom(testResults, domId) {
   const url = `${sessionUrl.origin}/api/images/dom/${domId}/?accountId=${accountId}&apiKey=${process.env.APPLITOOLS_API_KEY_READ}`
 
   const response = await fetch(url)
-  return response.json()
+  const dom = await response.json()
+
+  return {
+    ...dom,
+    getNodesByAttribute: attr => getNodesByAttribute(dom, attr),
+  }
+
+  function getNodesByAttribute(node, attr) {
+    const result = []
+    if (node.attributes && node.attributes[attr]) {
+      result.push(node)
+    }
+    if (node.childNodes) {
+      node.childNodes.forEach(node => result.push(...getNodesByAttribute(node, attr)))
+    }
+    return result
+  }
 }
 
 module.exports = {getTestInfo, getDom}