+++ packages/visual-grid-client/src/sdk/calculateSelectorsToFindRegionsFor.js
@@ -1,12 +1,12 @@
 'use strict'
 
+const {Region} = require('@applitools/eyes-sdk-core/shared')
+
 function selectorObject({selector, type}) {
   return type === 'xpath' || type === 'css' ? {type, selector} : selector
 }
 
 function calculateSelectorsToFindRegionsFor({
-  sizeMode,
-  selector,
   ignore,
   layout,
   strict,
@@ -14,8 +14,6 @@ function calculateSelectorsToFindRegionsFor({
   accessibility,
   floating,
 }) {
-  let selectorsToFindRegionsFor =
-    sizeMode === 'selector' || sizeMode === 'full-selector' ? [selector] : undefined
   const userRegions = Object.entries({
     ignore,
     layout,
@@ -31,58 +29,51 @@ function calculateSelectorsToFindRegionsFor({
   }, new Map())
 
   if (userRegions.size === 0) {
-    return {selectorsToFindRegionsFor, getMatchRegions}
+    return {getMatchRegions}
   }
 
-  const selectors = Array.from(userRegions.values()).reduce((prev, regions) => {
+  const selectorsToFindRegionsFor = Array.from(userRegions.values()).reduce((prev, regions) => {
     prev.push(...regions.filter(region => region.selector).map(selectorObject))
     return prev
   }, [])
 
   // NOTE: in rare cases there might be duplicates here. Intentionally not removing them because later we map `selectorsToFindRegionsFor` to `selectorRegions`.
-  return {
-    selectorsToFindRegionsFor: (selectorsToFindRegionsFor || []).concat(selectors),
-    getMatchRegions,
-  }
+  return {selectorsToFindRegionsFor, getMatchRegions}
 
-  function getMatchRegions({selectorRegions}) {
-    const imageLocationRegion =
-      sizeMode === 'selector' || sizeMode === 'full-selector' ? selectorRegions[0][0] : undefined
-    let selectorRegionIndex = imageLocationRegion ? 1 : 0
-    return Array.from(userRegions.entries()).reduce(
-      (allRegions, [regionName, userInput]) => {
-        const regionValues = userInput.reduce((regions, userRegion) => {
-          const codedRegions = userRegion.selector
-            ? selectorRegions[selectorRegionIndex++]
-            : [userRegion]
+  function getMatchRegions({selectorRegions, imageLocation}) {
+    let selectorRegionIndex = 0
+    return Array.from(userRegions.entries()).reduce((allRegions, [regionName, userInput]) => {
+      const regionValues = userInput.reduce((regions, userRegion) => {
+        const codedRegions = userRegion.selector
+          ? selectorRegions[selectorRegionIndex++]
+          : [userRegion]
 
-          if (codedRegions && codedRegions.length > 0) {
-            codedRegions.forEach(region => {
-              const regionObject = regionify({region, imageLocationRegion})
-              regions.push(regionWithUserInput({regionObject, userRegion, regionName}))
-            })
-          }
+        if (codedRegions && codedRegions.length > 0) {
+          codedRegions.forEach(region => {
+            const regionObject = regionify({region})
+            if (imageLocation && userRegion.selector) {
+              regionObject.left = Math.max(0, regionObject.left - imageLocation.x)
+              regionObject.top = Math.max(0, regionObject.top - imageLocation.y)
+            }
+            regions.push(regionWithUserInput({regionObject, userRegion, regionName}))
+          })
+        }
 
-          return regions
-        }, [])
-        allRegions[regionName] = regionValues
-        return allRegions
-      },
-      {imageLocationRegion},
-    )
+        return regions
+      }, [])
+      allRegions[regionName] = regionValues
+      return allRegions
+    }, {})
   }
 }
 
-function regionify({region, imageLocationRegion}) {
-  if (imageLocationRegion && region['getWidth']) {
-    return {
-      width: region.getWidth(),
-      height: region.getHeight(),
-      left: Math.max(0, region.getLeft() - imageLocationRegion.getLeft()),
-      top: Math.max(0, region.getTop() - imageLocationRegion.getTop()),
-    }
-  } else {
-    return region['toJSON'] ? region.toJSON() : region
+function regionify({region}) {
+  const regionInst = new Region(region)
+  return {
+    width: regionInst.getWidth(),
+    height: regionInst.getHeight(),
+    left: regionInst.getLeft(),
+    top: regionInst.getTop(),
   }
 }