+++ packages/eyes-sdk-core/CHANGELOG.md
@@ -3,6 +3,20 @@
 ## Unreleased
 
 
+## 12.11.0 - 2020/12/29
+
+- export `takeDomSnapshots` function for JS layout in Storybook
+- export `takeDomSnapshots` function for JS layout in Storybook
+## 12.10.2 - 2020/12/29
+
+- bring back verify-installed-versions
+## 12.10.1 - 2020/12/29
+
+- fix `extractText` method
+- add `imagePositionInActiveFrame` to `RenderStatusResults`
+- updated to @applitools/driver@1.0.2 (from 1.0.1)
+- updated to @applitools/screenshoter@1.0.2 (from 1.0.1)
+
 ## 12.10.0 - 2020/12/18
 
 - add new option to browser list ([Trello 693](https://trello.com/c/cEKupD0M))